Update build and watch scripts (using chokidar)
update: build script content moved into a module (generator.js) to
be used in a build and the chokidar-based watch scripts.

In 2019, chokidar was not watching file changes properly, thus the
npm script was named "watch-exp". The default watch script is
using nodemon. Both will be kept in the future.
