+++ lib/watch.js
@@ -1,6 +1,6 @@
 const chokidar = require('chokidar')
 const log = require('./utils/logger')
-const buildSite = require('./build-module')
+const buildSite = require('./generator')
 
 // Build site from ./src to ./public folder
 buildSite()
